+++ common/src/main/java/dev/qixils/crowdcontrol/common/command/impl/food/FoodSubCommand.java
@@ -22,7 +22,7 @@ public class FoodSubCommand<P> implements ImmediateCommand<P> {
 	public @NotNull Component getProcessedDisplayName(@NotNull Request request) {
 		if (request.getParameters() == null)
 			return getDefaultDisplayName();
-		int amount = (int) request.getParameters()[0];
+		int amount = (int) (double) request.getParameters()[0];
 		return getDefaultDisplayName().args(Component.text(amount));
 	}
 
@@ -31,7 +31,7 @@ public class FoodSubCommand<P> implements ImmediateCommand<P> {
 	public Response.Builder executeImmediately(@NotNull List<@NotNull P> players, @NotNull Request request) {
 		if (request.getParameters() == null)
 			return request.buildResponse().type(Response.ResultType.UNAVAILABLE).message("CC is improperly configured and failing to send parameters");
-		int amount = (int) request.getParameters()[0] * 2;
+		int amount = ((int) (double) request.getParameters()[0]) * 2;
 
 		Response.Builder resp = request.buildResponse().type(Response.ResultType.RETRY).message("Player's hunger is already empty");
 		for (P rawPlayer : players) {
